SA-1849 - Use usePrepareQuery hook for CompareByAggregatedMetrics component
